+++ BattleEngine/src/battleEngine/data/models/Armor.java
@@ -4,7 +4,7 @@ public class Armor {
     public String name;
     public double armorHP;
 
-    public Armor(String name, double armorHP) {
+    public Armor(String name, int armorHP) {
         this.name = name;
         this.armorHP = armorHP;
     }
